requirements.txt


numpy
pandas
scikit-learn
torch
torchvision


Flask
Flask-Cors
Werkzeug


SQLAlchemy
python-dotenv


opencv-python
Pillow

joblib
scikit-image

pytest

kaggle
fpdf

tqdm
requests